Change method for identifing doctor profile.
Profile for the doctor was identified using name; when that name changed
(i.e., by adding an "MD" at the end of it), the identification failed
and the returned data was null, causing the provider page to break.
This commit updates the query to work with the profile slug instead,
which should be more stable.